MAINT: allow __array_ufunc__ = None to force binops to defer.
In previous versions, one could force ndarray binops to defer by
setting a high __array_priority__.  With __array_ufunc__ this gets
ignored, and this commit ensures it is still possible to avoid
using the standard python language feature that setting something
to None means it is not implemented. In consequence, inside a
ufunc, if __array_ufunc__ is None, it will be treated as if it
had returned NotImplemented (leading to a TypeError if no other
object had a functioning __array_ufunc__ override).